Add files via upload
Filled in equations at the top. Prob would be good to double check my math. 

I updated pluto to be easier to work with by eliminating the need to give jupiter starting conditions. If we want jupiter to be a dynamical part of the system, I saved the original code to develop later at the bottom of the notebook.

I also wrote a basic simulator that can run many iterations of our class with random inputs and plot them. 

See and add to "todo" section at top.